# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Production API for Shiprion quotes, shipments, tracking, accounts, support, and operations.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments
      description: Admins see all shipments. Regular users see only their own.
      parameters:
        - {
            name: page,
            in: query,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentPage"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentBody"
      responses:
        "201":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get a shipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentBody"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteShipment
      tags: [shipments]
      summary: Delete a shipment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /shipments/{id}/assign:
    patch:
      operationId: assignShipmentCustomer
      tags: [shipments]
      summary: Assign a legacy shipment to an active customer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [customerId],
                    properties: { customerId: { type: integer } },
                  },
              },
          }
      responses:
        "200":
          {
            description: Shipment assigned,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Shipment" } },
              },
          }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        requestId:
          type: string
      required: [code, message]
...
    Shipment:
      type: object
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        senderId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        sourceQuoteId:
          type: ["integer", "null"]
        priceCents:
          type: ["integer", "null"]
        ownershipNeedsReview:
          type: boolean
        senderName:
          type: ["string", "null"]
        senderPhone:
          type: ["string", "null"]
        senderStreetAddress:
          type: ["string", "null"]
        senderHomeAddress:
          type: ["string", "null"]
        senderCity:
          type: ["string", "null"]
        senderPostalCode:
          type: ["string", "null"]
        recipientName:
          type: string
        recipientEmail:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        recipientStreetAddress:
          type: ["string", "null"]
        recipientHomeAddress:
          type: ["string", "null"]
        recipientCity:
          type: ["string", "null"]
        recipientPostalCode:
          type: ["string", "null"]
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
          enum:
            [
              registered,
              pending,
              picked_up,
              in_transit,
              customs,
              on_hold_customs,
              arrived_at_port,
              out_for_delivery,
              delivered,
              cancelled,
            ]
        weightKg:
          type: ["number", "null"]
        estimatedDelivery:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - trackingNumber
        - senderId
        - customerId
        - ownershipNeedsReview
        - recipientName
        - recipientEmail
        - origin
        - destination
        - status
        - weightKg
        - estimatedDelivery
        - createdAt
        - updatedAt
    ShipmentPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Shipment" } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
      required: [items, page, pageSize, total]
    CreateShipmentBody:
      type: object
      properties:
        customerId:
          type: integer
        senderName:
          type: ["string", "null"]
          maxLength: 100
        senderPhone:
          type: ["string", "null"]
          maxLength: 30
        senderAddress:
          type: ["string", "null"]
          maxLength: 200
        senderStreetAddress:
          type: ["string", "null"]
          maxLength: 200
        senderHomeAddress:
          type: ["string", "null"]
          maxLength: 200
        senderCity:
          type: ["string", "null"]
          maxLength: 100
        senderPostalCode:
          type: ["string", "null"]
          maxLength: 30
        recipientName:
          type: string
          minLength: 2
          maxLength: 100
        recipientEmail:
          type: ["string", "null"]
          format: email
          maxLength: 254
        recipientPhone:
          type: ["string", "null"]
          maxLength: 30
        recipientAddress:
          type: ["string", "null"]
          maxLength: 200
        recipientStreetAddress:
          type: ["string", "null"]
          maxLength: 200
        recipientHomeAddress:
          type: ["string", "null"]
          maxLength: 200
        recipientCity:
          type: ["string", "null"]
          maxLength: 100
        recipientPostalCode:
          type: ["string", "null"]
          maxLength: 30
        origin:
          type: string
          minLength: 2
          maxLength: 120
        destination:
          type: string
          minLength: 2
          maxLength: 120
        weightKg:
          type: ["number", "null"]
        estimatedDelivery:
          type: ["string", "null"]
          format: date
        adminNote:
          type: ["string", "null"]
          maxLength: 500
          description: Optional note from admin appended to the creation SMS
      required: [customerId, recipientName, origin, destination]
    UpdateShipmentBody:
      type: object
      properties:
        senderName:
          type: ["string", "null"]
          maxLength: 100
        senderPhone:
          type: ["string", "null"]
          maxLength: 30
        senderStreetAddress:
          type: ["string", "null"]
          maxLength: 200
        senderHomeAddress:
          type: ["string", "null"]
          maxLength: 200
        senderCity:
          type: ["string", "null"]
          maxLength: 100
        senderPostalCode:
          type: ["string", "null"]
          maxLength: 30
        recipientName:
          type: string
          minLength: 2
          maxLength: 100
        recipientEmail:
          type: ["string", "null"]
          format: email
          maxLength: 254
        recipientPhone:
          type: ["string", "null"]
          maxLength: 30
        recipientStreetAddress:
          type: ["string", "null"]
          maxLength: 200
        recipientHomeAddress:
          type: ["string", "null"]
          maxLength: 200
        recipientCity:
          type: ["string", "null"]
          maxLength: 100
        recipientPostalCode:
          type: ["string", "null"]
          maxLength: 30
        origin:
          type: string
          minLength: 2
          maxLength: 120
        destination:
          type: string
          minLength: 2
          maxLength: 120
        status:
          type: string
          enum:
            [
              registered,
              pending,
              picked_up,
              in_transit,
              customs,
              on_hold_customs,
              arrived_at_port,
              out_for_delivery,
              cancelled,
            ]
        weightKg:
          type: ["number", "null"]
        estimatedDelivery:
          type: ["string", "null"]
          format: date
        adminNote:
          type: ["string", "null"]
          maxLength: 500
          description: Optional note from admin appended to the status-change SMS
...